# openapi.yml
openapi: 3.1.0
info:
  title: SMSGang API
  version: 1.0.0
  description: |
    The SMSGang REST API enables purchasing temporary virtual phone numbers for
    SMS verification on 300+ global platforms. All monetary values are in **Nigerian Naira (₦)**.

    ## Authentication
    The API uses **Laravel Sanctum** bearer tokens. After registering or logging in,
    include the token in every protected request via the `Authorization` header:
    ```
    Authorization: Bearer {token}
    ```

    ## Rate Limits
    | Scope | Limit |
    |---|---|
    | `auth` | 10 req / min |
    | `api` | 60 req / min |
    | `buy` | 5 purchases / min |
    | `sms-check` | 10 checks / min |
    | `webhook` | 30 req / min |

  contact:
    name: SMSGang Support
  license:
    name: Proprietary
servers:
  - url: https://api.smsgang.ng/api
    description: Production
  - url: http://localhost:8000/api
    description: Local Development
...
paths:
...
  /services/{service}/countries:
    get:
      tags: [Catalog]
      summary: Get available countries and live prices for a service
      operationId: catalog.service.countries
      parameters:
        - $ref: '#/components/parameters/ServiceSlug'
      responses:
        '200':
          description: Countries with real-time prices
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceCountryPrice'
        '404':
          $ref: '#/components/responses/NotFound'
  /countries/{country}/services:
    get:
      tags: [Catalog]
      summary: Get available services for a country
      operationId: catalog.country.services
      parameters:
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Services available in the given country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCollection'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ACTIVATIONS ──────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum personal access token

  # ── PARAMETERS ──────────────────────────────
  parameters:
...
    ServiceSlug:
      name: service
      in: path
      required: true
      schema:
        type: string
      description: Service slug (e.g. `whatsapp`)
...
    CountryCode:
      name: country
      in: path
      required: true
      schema:
        type: string
      description: Country ISO code or slug (e.g. `NG`)
...
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: WhatsApp
        slug:
          type: string
          example: whatsapp
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    Country:
      type: object
      properties:
        id:
          type: integer
          example: 5
        name:
          type: string
          example: Nigeria
        code:
          type: string
          example: NG
        flag:
          type: string
          example: 🇳🇬
        is_active:
          type: boolean
          example: true
...
    ServiceCountryPrice:
      type: object
      description: Live price entry as returned by /services/{service}/countries
      properties:
        country:
          $ref: '#/components/schemas/Country'
        cost_ngn:
          type: number
          format: float
          description: Final price in NGN including exchange rate and global markup
          example: 350.00
        available_count:
          type: integer
          example: 4821
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ServiceCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Service'
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Not found.
...